# packages/contracts/openapi/course-authoring.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Learning MORE CourseAuthoring API",
    "version": "1"
  },
  "paths": {
    "/api/v1/outline-sessions": {
      "post": {
        "operationId": "createOutlineSession",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOutlineSessionBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OutlineSession created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutlineSessionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Application problem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationProblem"
                }
              }
            }
          },
          "404": {
            "description": "Application problem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationProblem"
                }
              }
            }
          },
          "409": {
            "description": "Application problem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationProblem"
                }
              }
            }
          },
          "412": {
            "description": "Application problem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationProblem"
                }
              }
            }
          },
          "428": {
            "description": "Application problem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationProblem"
                }
              }
            }
          },
          "500": {
            "description": "Application problem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationProblem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/outline-sessions/{sessionId}": {
      "get": {
        "operationId": "getOutlineSession",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OutlineSession view",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutlineSessionViewResponse"
                }
              }
            }
          },
          "400": {
            "description": "Application problem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationProblem"
                }
              }
            }
          },
          "404": {
            "description": "Application problem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationProblem"
                }
              }
            }
          },
          "409": {
            "description": "Application problem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationProblem"
                }
              }
            }
          },
          "412": {
            "description": "Application problem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationProblem"
                }
              }
            }
          },
          "428": {
            "description": "Application problem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationProblem"
                }
              }
            }
          },
          "500": {
            "description": "Application problem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationProblem"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteOutlineSessionDraft",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Unconfirmed outline session permanently deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteOutlineSessionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Application problem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationProblem"
                }
              }
            }
          },
          "404": {
            "description": "Application problem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationProblem"
                }
              }
            }
          },
          "409": {
            "description": "Application problem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationProblem"
                }
              }
            }
          },
          "412": {
            "description": "Application problem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationProblem"
                }
              }
            }
          },
          "428": {
            "description": "Application problem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationProblem"
                }
              }
            }
          },
          "500": {
            "description": "Application problem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationProblem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/outline-sessions/{sessionId}/messages": {
      "post": {
        "operationId": "appendOutlineSessionMessage",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppendOutlineSessionMessageBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Message accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutlineMessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Application problem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationProblem"
                }
              }
            }
          },
          "404": {
            "description": "Application problem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationProblem"
                }
              }
            }
          },
          "409": {
            "description": "Application problem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationProblem"
                }
              }
            }
          },
          "412": {
            "description": "Application problem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationProblem"
                }
              }
            }
          },
          "428": {
            "description": "Application problem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationProblem"
                }
              }
            }
          },
          "500": {
            "description": "Application problem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationProblem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/courses/{courseId}/outline-revisions": {
      "post": {
        "operationId": "reviseCourseOutline",
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviseCourseOutlineBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Outline revision created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutlineRevisionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Application problem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationProblem"
                }
              }
            }
          },
          "404": {
            "description": "Application problem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationProblem"
                }
              }
            }
          },
          "409": {
            "description": "Application problem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationProblem"
                }
              }
            }
          },
          "412": {
            "description": "Application problem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationProblem"
                }
              }
            }
          },
          "428": {
            "description": "Application problem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationProblem"
                }
              }
            }
          },
          "500": {
            "description": "Application problem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationProblem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateOutlineSessionBody": {
        "type": "object",
        "properties": {
          "topic": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          },
          "courseMode": {
            "type": "string",
            "enum": [
              "standard",
              "brainstorm",
              "argument_clash",
              "case_study",
              "business_insight",
              "process_decomposition",
              "decision_analysis",
              "cross_explore",
              "reading_seminar"
            ]
          }
        },
        "required": [
          "topic",
          "courseMode"
        ],
        "additionalProperties": false
      },
      "AppendOutlineSessionMessageBody": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100000
          }
        },
        "required": [
          "content"
        ],
        "additionalProperties": false
      },
...
      "ReviseCourseOutlineBody": {
        "type": "object",
        "properties": {
          "sourceCandidateVersionId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          }
        },
        "required": [
          "sourceCandidateVersionId"
        ],
        "additionalProperties": false
      },
...
      "DeleteOutlineSessionResponse": {
        "type": "object",
        "properties": {
          "outlineSessionId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          }
        },
        "required": [
          "outlineSessionId",
          "deletedAt"
        ],
        "additionalProperties": false
      },
...
    }
  }
}